+++ src/controllers/Customer/readCustomer.ts
@@ -25,3 +25,22 @@ export const getCustomer = async (req: Request, res: Response) => {
       .json({ success: false, error: 'Comuniquese con su administrador' })
   }
 }
+
+export const getCustomers = async (_req: Request, res: Response) => {
+  try {
+    const customers = await CustomerModel.find()
+
+    res.json({
+      success: true,
+      data: customers,
+    })
+  } catch (error) {
+    if (error instanceof Error) {
+      console.error(`Error to retrieve all customers: ${error.message}`)
+    }
+
+    res
+      .status(500)
+      .json({ success: false, error: 'Comuniquese con su administrador' })
+  }
+}
