+++ src/routes/customer.ts
@@ -1,20 +1,17 @@
 import { Router } from 'express'
 
-import {
-  deleteCustomer,
-  getCustomer,
-  getCustomers,
-  postCustomer,
-  putCustomer,
-} from '@controllers/customer'
 import { authToken } from '@src/middlewares/auth'
+import { createCustomer } from '@controllers/Customer/createCustomer'
+import { getCustomer, getCustomers } from '@controllers/Customer/readCustomer'
+import { deleteCustomer } from '@controllers/Customer/deleteCustomer'
+import { updateCustomer } from '@controllers/Customer/updateCustomer'
 
 const customerRouter = Router()
 
 customerRouter.get('/:id', getCustomer)
 customerRouter.get('/', authToken, getCustomers)
-customerRouter.post('/', postCustomer)
-customerRouter.put('/:id', putCustomer)
+customerRouter.post('/', createCustomer)
+customerRouter.put('/:id', updateCustomer)
 customerRouter.delete('/:id', deleteCustomer)
 
 export default customerRouter
